Stabilize full E2E suite to 19/19 green
- Add migration 028: closes_at column on lots (required for anti-snipe)
- Fix bidService.applyAntiSnipe: $3::int cast in jsonb_build_object
- Fix rateLimit: skip limiters in non-production to prevent test flakiness
- Fix bidding.spec.js cleanBids: remove invalid closes_at = closes_at no-op
- Add scripts/seed-test-fixtures.js for stable E2E fixture data

+++ src/services/bidService.js
@@ -173,7 +173,7 @@ async function applyAntiSnipe(client, lot) {
               jsonb_build_object(
                 'lot_id',        $1::text,
                 'closes_at',     $2::text,
-                'visible_cents', $3
+                'visible_cents', $3::int
               )
        FROM (
               SELECT user_id FROM bids       WHERE lot_id = $1
